s2e_env: moved server and tui out of commands module
