+++ s2e_env/server/cgc_interface.py
@@ -20,8 +20,8 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-import os
 import logging
+import os
 
 from . import coverage
 from .coverage import Coverage
